@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@giantmachines/redux-websocket": "^1.3.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
@@ -14,13 +13,15 @@
     "@types/react-router-dom": "^5.1.6",
     "@types/react-router-redux": "^5.0.18",
     "autobind-decorator": "^2.4.0",
+    "bootstrap": "^5.2.3",
     "react": "^17.0.1",
+    "react-bootstrap": "^2.7.0",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.4",
     "set": "https://github.com/GeneralSet/Set",
-    "typescript": "~3.7.2",
+    "typescript": "~3.8.2",
     "typestyle": "^2.1.0"
   },
   "scripts": {
