@@ -1,186 +1,154 @@
 import * as React from "react";
-import autobind from "autobind-decorator";
-import Board from "components/game/board";
-import PreviousSelection from "components/game/previousSelection";
-import { match } from "react-router-dom";
+import { useEffect, useState } from 'react';
+import { Board } from "components/game/board";
+import {PreviousSelection} from "components/game/previousSelection";
+import Modal from "react-bootstrap/Modal";
+import Button from "react-bootstrap/Button";
+import { useSelector } from "react-redux";
+import { ReduxState } from "reducers";
+
 const GeneralSet =
   process.env.NODE_ENV !== "test" ? import("set/pkg/set") : ({} as any);
 
-interface Props {
-  match: match<{ gameType: gameType }>;
-}
 
-interface State {
-  board: string[];
-  selected: string[];
-  hint: string[];
-  previousSelection: string[];
-  deck: string[];
-  alert: {
-    isError: boolean;
-    message: string;
-  };
-  points: number;
-  numberOfSets: number;
-}
 
-@autobind
-export default class Game extends React.Component<Props, State> {
-  private set: any;
-  private readonly cardsForSet = 3;
+const Game = () => {
+  const [set, setSet] = useState<any>();
+  const [points, setPoints] = useState<number>(0);
+  const [hint, setHint] = useState<string[]>([]);
+  const [selected, setSelected] = useState<string[]>([]);
+  const [previousSelection, setPreviousSelection] = useState<string[]>([]);
+  const [message, setMessage] = useState<string>("");
+  const deckData = useSelector(
+    (state: ReduxState) => state.singlePlayer.deck
+  );
+  const features = deckData.features.length;
+  const options = deckData.numOptions;
+  const boardSize = features * options;
 
-  constructor(props: Props) {
-    super(props);
+  useEffect(() => {
     GeneralSet.then((s: any) => {
-      this.set = s.Set.new(
-        this.props.match.params.gameType === "custom" ? 3 : 4,
-        3,
-        this.props.match.params.gameType === "custom" ? 9 : 12
-      );
-      this.setState({
-        deck: this.set.get_deck().split(","),
-        board: this.set.get_board().split(","),
-        selected: [],
-        hint: [],
-        previousSelection: [],
-        alert: {
-          isError: false,
-          message: "",
-        },
-        numberOfSets: this.set.sets,
-        points: 0,
-      });
+      setSet(s.Set.new(
+        features,
+        options,
+        boardSize
+      ));
     });
-    this.state = {
-      deck: [],
-      board: [],
-      selected: [],
-      hint: [],
-      previousSelection: [],
-      alert: {
-        isError: false,
-        message: "",
-      },
-      numberOfSets: 0,
-      points: 0,
-    };
+  },[features, options, boardSize]);
+
+  const restartGame = () => {
+    GeneralSet.then((s: any) => {
+      setSet(s.Set.new(
+        features,
+        options,
+        boardSize
+      ));
+    });
+    setPreviousSelection([]);
+    setPoints(0);
+  };
+
+  if (!set) {
+    return (
+      <div>Loading...</div>
+    )
   }
 
-  selectCard(id: string, selectedIndex: number) {
-    const board = this.state.board;
-    const selected = this.state.selected;
+  const deck = set.get_deck() ? set.get_deck().split(",") : [];
+  const board = set.get_board() ? set.get_board().split(",") : [];
+  const numberOfSets = set.sets;
+
+  const selectCard = (id: string) => {
+    const selectedClone = [...selected];
 
     // update selected cards
-    if (selected.includes(id)) {
-      selected.forEach((cardId: string, index: number) => {
+    if (selectedClone.includes(id)) {
+      selectedClone.forEach((cardId: string, index: number) => {
         if (cardId === id) {
-          selected.splice(index, 1);
+          selectedClone.splice(index, 1);
         }
       });
     } else {
-      selected.push(id);
+      selectedClone.push(id);
+    }
+    if (selectedClone.length < options) {
+      setSelected(selectedClone);
+      return;
     }
 
-    // update board
-    if (selected.length >= this.cardsForSet) {
-      this.verifySet();
-    } else {
-      this.setState({ board, selected });
-    }
-  }
-
-  clearSelection() {
-    this.setState({ selected: [] });
-  }
-
-  verifySet(): boolean {
-    const selected = this.state.selected;
-    this.clearSelection();
-    // ensure right num cards selected
-    if (selected.length < this.cardsForSet) {
-      this.setState({
-        alert: { isError: true, message: "Error: not enough cards selected" },
-        previousSelection: selected,
-      });
-      return false;
-    }
-
+    setHint([]);
+    setSelected([]);
+    setPreviousSelection(selectedClone);
     // check for set
-    const isValidSet = this.set.is_set(selected.join(","));
+    const isValidSet = set.is_set(selectedClone.join(","));
     if (!isValidSet) {
-      this.setState({
-        alert: { isError: true, message: `-1 Not a set.` },
-        points: this.state.points - 1,
-        previousSelection: selected,
-      });
-      return false;
+      setPoints(points - 1);
+      setMessage("-1 Not a set.");
+      return;
     }
 
     // Set found
-    this.set = this.set.update_board(selected.join(","));
-    this.setState({
-      alert: { isError: false, message: "+1 Set!" },
-      points: this.state.points + 1,
-      board: this.set.get_board().split(","),
-      deck: this.set.get_deck() ? this.set.get_deck().split(",") : [],
-      numberOfSets: this.set.sets,
-      previousSelection: selected,
-      hint: [],
-    });
-    return true;
+    setSet(set.update_board(selectedClone.join(",")));
+    setPoints(points + 1);
+    setMessage("+1 Set!");
   }
 
-  private hint(_event: React.MouseEvent<HTMLButtonElement>): void {
-    this.setState({
-      hint: this.set.hint(this.state.board.join(",")).split(","),
-    });
+  const giveHint = (_event: React.MouseEvent<HTMLButtonElement>) => {
+    setHint(set.hint(board.join(",")).split(","))
   }
 
-  render() {
-    return (
-      <div>
-        <p>is end? {this.set && this.set.is_end() ? "yes" : "no"}</p>
-        <div className="container mb-2">
-          <div className="row">
-            <div className="col-sm">
-              <table className="table table-borderless w-auto">
-                <tbody>
-                  <tr>
-                    <td className="text-left">Points</td>
-                    <td>{this.state.points}</td>
-                  </tr>
-                  <tr>
-                    <td className="text-left">Cards Remaining</td>
-                    <td>{this.state.deck.length}</td>
-                  </tr>
-                  <tr>
-                    <td className="text-left">Sets on the Board</td>
-                    <td>{this.state.numberOfSets}</td>
-                  </tr>
-                </tbody>
-              </table>
-            </div>
-            <div className="col-sm text-right">
-              <button onClick={this.hint} className="btn btn-primary btn-sm">
-                Hint
-              </button>
-              <PreviousSelection
-                cards={this.state.previousSelection}
-                gameType={this.props.match.params.gameType}
-                message={this.state.alert.message}
-                success={!this.state.alert.isError}
-              />
-            </div>
+  return (
+    <div>
+      <Modal show={set && set.is_end()}>
+        <Modal.Header>
+          <Modal.Title>Game Over</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          Score: {points}
+        </Modal.Body>
+        <Modal.Footer>
+          <Button onClick={restartGame}>Play Again</Button>
+        </Modal.Footer>
+      </Modal>
+      <button onClick={giveHint} className="btn btn-secondary btn-sm">
+        Hint
+      </button>
+      <div className="container mb-2">
+        <div className="row">
+          <div className="col-sm">
+            <table className="table table-borderless w-auto" style={{color: "white"}}>
+              <tbody>
+                <tr>
+                  <td style={{textAlign: "left"}}>Score</td>
+                  <td>{points}</td>
+                </tr>
+                <tr>
+                  <td style={{textAlign: "left"}}>Cards Remaining</td>
+                  <td>{deck.length}</td>
+                </tr>
+                <tr>
+                  <td style={{textAlign: "left"}}>Sets on the Board</td>
+                  <td>{numberOfSets}</td>
+                </tr>
+              </tbody>
+            </table>
+          </div>
+          <div className="col-sm">
+            <PreviousSelection
+              cards={previousSelection}
+              message={message}
+            />
           </div>
         </div>
-        <Board
-          board={this.state.board}
-          selected={this.state.selected}
-          hint={this.state.hint}
-          gameType={this.props.match.params.gameType}
-          onSelect={this.selectCard}
-        />
       </div>
-    );
-  }
+      <Board
+        board={board}
+        selected={selected}
+        hint={hint}
+        onSelect={selectCard}
+      />
+    </div>
+  );
 }
+
+export default Game;
