@@ -4,6 +4,6 @@ import App from "./App";
 
 test("renders learn react link", () => {
   const { getByText } = render(<App />);
-  const linkElement = getByText(/Single Player/i);
+  const linkElement = getByText(/Play/i);
   expect(linkElement).toBeInTheDocument();
 });
