@@ -4,10 +4,10 @@ import { Link } from "react-router-dom";
 const Layout: React.FC = (props) => (
   <div
     className="text-center"
-    style={{ backgroundColor: "#A79ACC", height: "100%" }}
+    style={{ backgroundColor: "#4A3665", height: "100%" }}
   >
     <h1 style={{ padding: "1rem 0" }}>
-      <Link to="/" style={{ color: "#513e96", textDecoration: "none" }}>
+      <Link to="/" style={{ color: "#fff", textDecoration: "none" }}>
         GeneralSet.io
       </Link>
     </h1>
