@@ -1,41 +1,33 @@
 import * as React from "react";
-import autobind from "autobind-decorator";
+import { useSelector } from "react-redux";
+import { ReduxState } from "reducers";
 import Card from "components/game/card";
 import "./index.css";
 
 interface Props {
   cards: string[];
-  gameType: gameType;
   message: string;
-  success: boolean;
 }
 
-@autobind
-export default class PreviousSelection extends React.Component<Props, {}> {
-  private cardDisplay(card: string, index: number): JSX.Element | null {
-    return (
-      <Card
-        key={index}
-        features={card}
-        selected={false}
-        gameType={this.props.gameType}
-      />
-    );
+export const PreviousSelection = ({cards, message}: Props) => {
+  const deck = useSelector(
+    (state: ReduxState) => state.singlePlayer.deck
+  );
+  if (cards.length <= 0) {
+    return null;
   }
-
-  render() {
-    if (this.props.cards.length <= 0) {
-      return null;
-    }
-    return (
-      <div
-        className={`previous-selection ${
-          this.props.success ? "success" : "error"
-        }`}
-      >
-        <div className="message text-left">{this.props.message}</div>
-        <div className="cards">{this.props.cards.map(this.cardDisplay)}</div>
+  return (
+    <div className="previous-selection">
+      <div style={{color: "white"}}>{message}</div>
+      <div className="cards">
+        {cards.map((card, index) => (
+          <Card
+            key={index}
+            selected={false}
+            svg={deck.cards[card]}
+          />
+        ))}
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
