@@ -1,5 +1,5 @@
 import { createStore, applyMiddleware } from 'redux';
-import websocket from '@giantmachines/redux-websocket';
+// import websocket from '@giantmachines/redux-websocket';
 import { rootReducer } from './reducers';
 
-export const store = applyMiddleware(websocket)(createStore)(rootReducer);
+export const store = applyMiddleware()(createStore)(rootReducer);
